+++ awslimitchecker/services/ec2.py
@@ -134,10 +134,6 @@ class _Ec2Service(_AwsService):
     def _find_usage_spot_instances(self):
         """calculate spot instance request usage and update Limits"""
         logger.debug('Getting spot instance request usage')
-        logger.warning("EC2 spot instance support is experimental and results "
-                       "may not me accurate in all cases. Please see the notes "
-                       "at: <http://awslimitchecker.readthedocs.io/en/latest"
-                       "/limits.html#ec2>")
         try:
             res = self.conn.describe_spot_instance_requests()
         except botocore.exceptions.ClientError as e:
@@ -153,10 +149,6 @@ class _Ec2Service(_AwsService):
             else:
                 logger.debug('NOT counting spot instance request %s state=%s',
                              req['SpotInstanceRequestId'], req['State'])
-        logger.debug('Setting "Max spot instance requests per region" limit '
-                     '(%s) current usage to: %d',
-                     self.limits['Max spot instance requests per region'],
-                     count)
         self.limits['Max spot instance requests per region']._add_current_usage(
             count,
             aws_type='AWS::EC2::SpotInstanceRequest'
@@ -175,8 +167,6 @@ class _Ec2Service(_AwsService):
             logger.error('Error: describe_spot_fleet_requests() response '
                          'includes pagination token, but pagination not '
                          'configured in awslimitchecker.')
-        # @TODO: assumption: 'Max target capacity for all spot fleets in region'
-        # only counts active fleets (not submitted)
         active_fleets = 0
         total_target_cap = 0
         lim_cap_per_fleet = self.limits['Max target capacity per spot fleet']
@@ -193,16 +183,11 @@ class _Ec2Service(_AwsService):
             launch_specs = len(
                 fleet['SpotFleetRequestConfig']['LaunchSpecifications'])
             total_target_cap += cap
-            logger.debug('Active fleet %s: target capacity=%s, %d launch specs',
-                         _id, cap,
-                         launch_specs)
             lim_cap_per_fleet._add_current_usage(
                 cap, resource_id=_id, aws_type='AWS::EC2::SpotFleetRequest')
             lim_launch_specs._add_current_usage(
                 launch_specs, resource_id=_id,
                 aws_type='AWS::EC2::SpotFleetRequest')
-        logger.debug('Total active spot fleets: %d; total target capacity '
-                     'for all spot fleets: %d', active_fleets, total_target_cap)
         self.limits['Max active spot fleets per region']._add_current_usage(
             active_fleets, aws_type='AWS::EC2::SpotFleetRequest'
         )